// src/models/regions/locations/spawn/traits/settlements/environment/types.ts
export type settlement__environment =
  | 'artistic patronage'
  | 'brilliant innovation'
  | 'corrupt laws'
  | 'crackdown'
  | 'decadent locals'
  | 'despotic lords'
  | 'faded remnant'
  | 'fallen prosperity'
  | 'hazardous district'
  | 'heavy fortifications'
  | 'hidden ruler'
  | 'imperious architect'
  | 'impoverished district'
  | 'incompetent leaders'
  | 'infamous prison'
  | 'local festival'
  | 'neglectful ruler'
  | 'new industry'
  | 'pilgrimage site'
  | 'punishment post'
  | 'regency council'
  | 'ritual combat'
  | 'ruins beneath'
  | 'trade hub'
  | 'unique product'
  | 'xenophobic locals'